+++ src/algorithm/bfs/openLock.ts
@@ -3,13 +3,13 @@
 export function openLock(dead: string[], target: string): number {
   const queue: string[] = [];
   const visited = new Set<string>();
-  const deadMap = new Set<string>();
+  const deadSet = new Set<string>();
 
   for (let i = 0; i < dead.length; i++) {
-    deadMap.add(dead[i]);
+    deadSet.add(dead[i]);
   }
 
-  if (deadMap.has('0000') === false) {
+  if (deadSet.has('0000') === false) {
     queue.push('0000');
     visited.add('0000');
   }
@@ -30,13 +30,13 @@ export function openLock(dead: string[], target: string): number {
         const tail = currentNode.slice(j + 1);
 
         const up = `${head}${(Number(currentNode[j]) + 1) % 10}${tail}`;
-        if (deadMap.has(up) === false && visited.has(up) === false) {
+        if (deadSet.has(up) === false && visited.has(up) === false) {
           queue.push(up);
           visited.add(up);
         }
 
         const bottom = `${head}${(Number(currentNode[j]) + 9) % 10}${tail}`;
-        if (deadMap.has(bottom) === false && visited.has(bottom) === false) {
+        if (deadSet.has(bottom) === false && visited.has(bottom) === false) {
           queue.push(bottom);
           visited.add(bottom);
         }